Scroll down loading for user collection and volume change operations
appendLocalCollection method now loads first n items to fill the screen.

appendLocalCollectionScroll method is called if viewport is scrolled
down and loads the next n item if screen is to hit bottom of viewport.

songCollection.itemCount is number of items to fill the screen. This var
is changed if screen height changes.

Volume icon changes according to volume level. Clicking on volume icon
mutes the sound, and clicking again sets the old volume level.